+++ freemarker-generator-cli/src/main/java/org/apache/freemarker/generator/cli/impl/ConfigurationSupplier.java
@@ -24,14 +24,18 @@ import freemarker.template.TemplateExceptionHandler;
 import freemarker.template.Version;
 import org.apache.freemarker.generator.cli.model.Settings;
 
+import java.util.Properties;
 import java.util.function.Supplier;
 
 import static freemarker.core.TemplateClassResolver.ALLOWS_NOTHING_RESOLVER;
 import static freemarker.template.Configuration.VERSION_2_3_29;
 import static java.util.Objects.requireNonNull;
+import static org.apache.freemarker.generator.base.util.PropertiesTransformer.filterKeyPrefix;
+import static org.apache.freemarker.generator.base.util.PropertiesTransformer.removeKeyPrefix;
 
 public class ConfigurationSupplier implements Supplier<Configuration> {
 
+    private static final String FREEMARGER_CONFIGURATION_SETTING_PREFIX = "freemarker.configuration.setting.";
     private static final Version FREEMARKER_VERSION = VERSION_2_3_29;
 
     private final Settings settings;
@@ -44,21 +48,38 @@ public class ConfigurationSupplier implements Supplier<Configuration> {
 
     @Override
     public Configuration get() {
-        final Configuration configuration = new Configuration(FREEMARKER_VERSION);
-        configuration.setAPIBuiltinEnabled(false);
-        configuration.setDefaultEncoding(settings.getTemplateEncoding().name());
-        configuration.setLocale(settings.getLocale());
-        configuration.setObjectWrapper(objectWrapper());
-        configuration.setOutputEncoding(settings.getOutputEncoding().name());
-        configuration.setTemplateLoader(templateLoader.get());
-        configuration.setTemplateExceptionHandler(TemplateExceptionHandler.RETHROW_HANDLER);
-        configuration.setNewBuiltinClassResolver(ALLOWS_NOTHING_RESOLVER);
-        return configuration;
+        try {
+            final Configuration configuration = new Configuration(FREEMARKER_VERSION);
+
+            // safe default configuration
+            configuration.setAPIBuiltinEnabled(false);
+            configuration.setNewBuiltinClassResolver(ALLOWS_NOTHING_RESOLVER);
+
+            // apply all "freemarker.configuration.setting" values
+            configuration.setSettings(freeMarkerConfigurationSettings());
+
+            // override current configuration with caller-provided settings
+            configuration.setDefaultEncoding(settings.getTemplateEncoding().name());
+            configuration.setLocale(settings.getLocale());
+            configuration.setObjectWrapper(objectWrapper());
+            configuration.setOutputEncoding(settings.getOutputEncoding().name());
+            configuration.setTemplateLoader(templateLoader.get());
+
+            return configuration;
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to create FreeMarker configuration", e);
+        }
     }
 
     private static DefaultObjectWrapper objectWrapper() {
         final DefaultObjectWrapperBuilder builder = new DefaultObjectWrapperBuilder(FREEMARKER_VERSION);
         builder.setIterableSupport(false);
         return builder.build();
+    }
+
+    private Properties freeMarkerConfigurationSettings() {
+        return removeKeyPrefix(
+                filterKeyPrefix(settings.getConfiguration(), FREEMARGER_CONFIGURATION_SETTING_PREFIX),
+                FREEMARGER_CONFIGURATION_SETTING_PREFIX);
     }
 }